backend/requirements.txt
# =============================================================================
# Core Django
# =============================================================================
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# =============================================================================
# Authentication
# =============================================================================
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1

# =============================================================================
# Database
# =============================================================================
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# =============================================================================
# Caching & Task Queue
# =============================================================================
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# =============================================================================
# API & Serialization
# =============================================================================
drf-spectacular==0.27.1
drf-nested-routers==0.94.1

# =============================================================================
# File handling
# =============================================================================
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34

# =============================================================================
# Utilities
# =============================================================================
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.31.0

# =============================================================================
# Financial / Currency
# =============================================================================
py-moneyed==3.0
django-money==3.4.1
babel==2.14.0

# =============================================================================
# PDF Generation
# =============================================================================
weasyprint==61.2
django-weasyprint==2.3.0

# =============================================================================
# Audit Trail
# =============================================================================
django-auditlog==2.3.0
django-simple-history==3.4.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.1
flake8==7.0.0
black==24.1.1
isort==5.13.2
